Reorganization of data in the frontend template (#41)
Hi,
in the future, I plan that all data interfaces should directly created
in data.tsx, and all random data in RandomData.tsx.

For this, I created new interfaces, deleted the old one in the Mission
table, and also did some changes in the mission table, because
MissionData now contains a mission_id field, which is of type number, to
make the fetching process in the detail view easier.

Random data now contains random data for the details view, which is
already finished, but currently, not a part of this PR.

First point of issue #39