Moved frontend documentation and added production server (#53)
Hi,
i found out that the long loading times in the frontend originate from
the dev environment (i. e.: `npm run dev`), but this enables features
like debugging and hot reload. When switching to production, the page
loads instantly.

To test it, type:
`npm run build`
and
`npm start`
in the frontend folder. Click on the URL.

This "resolves" issue #43 by extending (and moving) the frontend
documentation.