Added basic dataset view (#56)
Hi, this PR is an extension of the detail view. It addresses issue #26
task "Make table entries clickable".

Also added a color effect when sliding with the mouse over the table. 

![image](https://github.com/user-attachments/assets/46cb66fa-46ae-4335-a6d3-bf255a0b5c28)

By clicking on a row, a new view opens, the so called "DatasetView".
This will display more details about a dataset in future PRs.

![image](https://github.com/user-attachments/assets/eb551fcf-7a37-4930-9578-6eb1d78be6db)