Fixed icon not found error (#65)
Hi,
Sebastian mentioned today, that there's an issue with the icon, in some
specific cases, it is still looking for an icon named "favicon.ico",
independently from our name changes in root.tsx and AppShell.tsx.

A few hours ago, I was able to reproduce the same error in a new detail
view implementation and realized, this has something todo with vite.
Renaming the icon back to favicon.ico fixes the issue. Unfortunately, I
don't came up with a reproducible test case, so, the error does not
happen all the time.

But you as a reviewer can clearly see, that I just changed the file
names. The icon is still the poly bot icon.

Issue #64