Add a Tag- and Colorpicker to the frontend (#72)
resolves #44 

# Changes
## frontend/app/utilities/TagPicker.tsx

![grafik](https://github.com/user-attachments/assets/40d11c34-e09f-4abc-942e-c4b31a4f4628)

I implemented a TagPicker with the following 3 functionalities:
- You can add new tags and choose a color for them (type in your tag
name, click the color you like and then click on the "+" or press enter)
The chosen color gets marked with a black box:

![grafik](https://github.com/user-attachments/assets/14d8cbd4-2c81-4fd9-9f72-83a97a6d6a4d)

- You can remove existing tags (click on the trashcan next to the tag)
- You can change the color of existing tags (click on the color palette
next to the tag)

![grafik](https://github.com/user-attachments/assets/a111cfa6-6bb3-43e9-a649-85dc2cd43a16)


These functionalities can be handled by the backend in the TagPicker
call with onAddTag, onRemoveTag, onChangeTag
I predefined 8 colors I like in the code. If desired we could also add a
custom color picker or change the predefined colors.

## frontend/app/RandomData.tsx
I added my predefined colors in the example Data for consistency

## fronend/app/pages/missions/Overview.tsx
I added a call of my TagPicker in the OverviewPage. The TagPicker now
gets opened when clicking on the tags in the row.
Now the call just needs to be connected to the backend