Frontend deep links details (#74)
Hi, this PR implements the second task of #63 
DetailView is now a separate page, meaning it can be accessed via link.

Run it, and click on a row in the main mission table, a new page will
open, back button will work then.

But should I still use the app shell? meaning the navbar and the title
bar at the top?