Fixed search bug (#86)
This fixes the search bug which resulted in less results. Another issue
resulting in an exception when no results are shown was also resolved.

renderedData (sortedData before) was fetched only once and was
overridden with each search, meaning the searchable space reduces over
time. This is addressed by introducing a new ground truth "fetchedData".

Issue #75