fetch API for tags (#98)
Resolves the first 3 Tasks of Issue #55

# Changes
- I added the following functions to the fetchapi
   - getTags()
   - addTagToMission()
   - removeTagFromMission(missionId, tagName)
   - changeTagColor(tagName, newColor)
   - createTag(tagName, color)
   - deleteTag(tagName)
   - getTagsByMission(missionID)
   - getMissionsByTag(tagName)
- I updated the fetchapi documentation with my new functions
- I used the fetchapi tag functions in the overview tagPicker
- if a tag gets added the fetchapi creates a new tag and adds it to the
mission
- if a tag gets removed from a mission the relations table entry gets
deleted. And if the tag isn't associated with another mission the tag
gets also deleted
   - on color change the fetch api changes the color of the tag

Now there needs to be a feature to add tags to a mission that already
exist. but that is something for a future PR.

# Testing
You can check if the database table gets updated correctly with your
clicks on the frontend.

# Note
I changed the flag to not use random data anymore. I think we can start
developing with the real data. But we can also leave it on random data
if you think thats better on the main branch