Bugfix: urlencode tag name (#114)
As we decided at our meeting the tag_name should be encoded in the url,
but this was already the case. I found out that django does decode any
encoded characters in the url and it does that, because apache2 does
that too.\
So even If there was a way to change this in the webserver provided by
django for development it wouldn't work in production when we have to
use a real webserver like apache2.\
This is a problem because then we can't distinguish between the url
`/restapi/tags/create/` and the url `/restapi/tags/<str:tag_name>` when
`tag_name` is `create/`.\
A workaround for this is, to encode the name in the url twice and let it
be decoded by django/apache2 once and decode once ourself.\
I implemented the needed changes in the backend and the frontend and
added some tests that specifically test the behavior of the REST API
with a tag named `create/`.

resolves #100 

## How to test:
run all tests with
```bash
python manage.py test
```
and manually test the frontend by 
1. starting the frontend+backend 
2. creating a tag with the name  `create/`
3. try changing the color of the tag
4. try creating new tags
5. try deleting the tag


When deleting the tag or changing the color you should see a request in
the output of `./manage.py runserver` with a url like this:
`/restapi/tags/create%252F`\
`%252F` is the `/` character that got encoded twice.