Finished implementation of login system (#106)
Hi,
this PR finishes the last task of issue #103 and therefore the login
system.

Quick overview:

![image](https://github.com/user-attachments/assets/10fbd214-1e1d-4b43-97f6-009cab7e130b)
There is currently no backend connection, so, i created a testuser:

```
username: admin
password: admin
```

When using invalid credentials, the UI shows an error:

![image](https://github.com/user-attachments/assets/2672401c-6c5c-4586-9ad4-7bcd09af9af4)

In `utilities/LoginHandler.tsx`, is the logic for handling usernames and
passwords.

On login, credentials are checked, if they are "right", meaning
`authentificator.use` returns a user, a session cookie is created. This
is checked for every page, and in case there is no or an invalid
session, the browser automatically redirects to the login page.

Credential checks and checks if the session is valid, only happens on
the server side.

If you want to read more about this:
https://remix.run/resources/remix-auth

Also the default route `/` is handled appropriately. 

Documentation is also updated.