UI searchable tags (#123)
resolves #46

# Changes
- I changed the sortData and the filterData function to also handle the
tagSearch.
- I changed the way tags get displayed and what happens when you click
on them
- When clicking a Tag, only the rows with this tag get displayed and the
clicked tag gets marked. when clicking more tags, the rows get filtered
for all rows which contain all the clicked tags. This implements the
"and" search functionality for tags.

![grafik](https://github.com/user-attachments/assets/63d4f23c-d2eb-47f8-aa66-fe2663d23164)

![grafik](https://github.com/user-attachments/assets/3f7c9388-c42d-4e3a-ae12-a8421b339dd8)

![grafik](https://github.com/user-attachments/assets/83bc713f-b4e2-4812-9dbf-186c54aa3743)
For this to work. I had to change the way the tag Picker gets opened in
the overview. It now gets opened by the little pencil Button next to the
tags



# little fixes
- I removed the redundant tagId from the tag interface