adding django tests to github actions (#130)
Hey there,

in this PR I added
[`django.yml`](https://github.com/brendel-group/mission_db/blob/8376c9285ec7646e08dd9a275c979d8d337a8e0d/.github/workflows/django.yml)
in our `workflows` folder. With this addition, we added our django
tests, which we usually run with`python manage.py test restapi` to our
GitHub Actions. This way, the tests automatically run with every push
and PR, just like we did with our `ruff` checks.

This resolves #96