Added frontend logic to change notes and location in detail view (#144)
Hi,
this PR resolves the third task of issue #135 by adding an ui to handle
notes and location changes in the detail view:

1) States was renamed to Information

![image](https://github.com/user-attachments/assets/93275549-6bcf-4a0b-9f32-ce7860cd1369)

By clicking a new UI appears:

![image](https://github.com/user-attachments/assets/d49b746b-e5b0-42cd-8045-8e8d0eb3df55)

Currently the maximum length is 42. Backend connection will happen in a
separate PR.

It is possible to get in a new line by pressing `shift+ enter`. Only
pressing enter triggers the update button.