CLI interactive mode improvements (#155)
I made some improvements for the interactive mode and the CLI in
general:
1. When the output is a table it uses Unicode characters for the lines
in the table. This behavior is now optional and can be disabled with the
environment variable `USE_UNICODE` that can be set to `False` in the
.env file. It then uses ASCII characters.
2. The table now correctly displays strings that contain newlines
(`\n`). It now prints the second or more lines of such a string in a new
line that contains only empty fields or the respective content.
3. The interactive mode now supports multi line input when using quotes.
Previously it was only possible to have multi line strings when using
arguments and not the interactive mode.
4. When an exception is thrown when using the interactive mode it is
printed like normal and the interactive mode is terminated correctly
(writing the history to file and potentially more)

I added documentation for the environment variable `USE_UNICODE` and the
multi line input.\
And I added a test for the output of multi line strings in the table.

I added the option to disable the usage of Unicode characters, because
when using the pager added in #151 on windows with setting `PAGER=less`
it doesn't support Unicode characters and prints something unreadable.

## How to test
Use the CLI with the interactive mode and try adding a mission that
contains more than one line in the name or notes field.\
Then test the output using `mission list`.
Set `USE_UNICODE=False` in the backend/backend/.env file and test the
output again.