Make the CLI use a pager (#151)
When the CLI prints a table it can happen that the table is too wide for
the terminal window and the formatting is messed up and basically
unusable.\
To fix this I used to pipe the output of the cli to `less` where I could
use the option `-S` and the table gets displayed correctly and I can
scroll up, down, left and right with the arrow keys and search for
words. Like this:
```bash
./cli.py mission list | less -S
```

With the interactive mode this won't work.\
So I made it that when a table is too wide it automatically gets
displayed by a pager that is selected by the `PAGER` or `MANPAGER`
environment variable or a default pager.\
For Linux the default pager in most cases is `less`.\
When your pager is `less` you can type in `-S` after it opened to
disable long line folding to correctly see the table.

For Windows there is no default pager and it just prints the table, but
you can set the `PAGER` environment variable to any pager you like.

## How to test:
1. Create for example a mission with a name and notes that is too long
to be displayed in one line on your terminal.
2. use `python cli.py mission list` or the interactive mode to display
the table in a pager or print it.