CLI Refactoring (#161)
The file cli.py was getting too big, so I split it up into multiple
files.

To do this I created an abstract class `Command` that has all the
abstract methods needed for implementing a command and made it that
every .py file in the folder cli_commands is imported by cli.py. Like
this all classes inheriting/implementing the abstract class are found
and automatically added to a dict of commands. Using this dict every
command is automatically added and no changes in cli.py are required
anymore to implement a new command.

Then I moved every existing command in its own file in the cli_commands
folder and made the required changes.

Most of this was copy+paste, so this PR may look very large, but the
only real implementation happens in `cli.py`, `cli_commands/__init__.py`
and `cli_commands/Command.py`

I also added some documentation to the end of the [cli
documentation](https://github.com/brendel-group/mission_db/blob/feature/cli-refactoring/docs/cli/README.md)
and moved all tests in separate files for each command.

This makes implementing new commands much easier and it reduces the
amount of merge conflicts, because different commands are implemented in
different files.

The print_table function is part of the abstract `Command` class so it
is available for every command.

## How to test:
This is a pure refactoring and no new functionality is added, so just
check for potentially introduced errors.