Tag picker edit tag menu (#163)
resolves #162 

# Changes
This PR implements a little menu in the tag picker to change the name
and color of an existing tag. This menu replaces the little color
palette icon that was used to change the color before.
The new edit tag menu can be opened by clicking on a tag that gets
listed in the tag picker

![grafik](https://github.com/user-attachments/assets/70f56847-2e93-4413-8bfe-9c9afdf7cb2a)
You can change the name and the color there. Once something was changed,
the 'Save' button can be clicked.

![grafik](https://github.com/user-attachments/assets/a54efa24-99f5-4d42-8abd-ebf6b727065b)
After clicking the 'Save' button, the changes get saved and the menu
closes. Alternatively you can also do that by pressing enter on your
keyboard.
There are 3 error cases that get checked once enter or 'Save' was
pressed:
- tag name already exists
- tag name is too long
- invaild color

I also disabled the 'Add existing tags' and the 'Remove all tags'
buttons if there are no already existing tags or no tags to delete.

![grafik](https://github.com/user-attachments/assets/20cec7f6-9adc-4262-b6b6-29e3a527cda1)


![grafik](https://github.com/user-attachments/assets/bbd46787-3ee0-4f18-b680-bc248885e404)