Rest API Session Authentication (#168)
I replaced the authentication via API-KEY by Session Authentication.
This means that no configuration except creating a user is required
anymore for authentication. When you login at the frontend a `sessionid`
and `csrftoken` cookie are created and used to authenticate you on the
backend.

For some reason the cookies don't get sent to the backend with the fetch
requests when I open the frontend with `localhost` but it works with
`127.0.0.1` (which is basically the same) I think that is something
browser specific and is only relevant for local testing. So from now on
just open the frontend with
[http://127.0.0.1:3000/](http://127.0.0.1:3000) for local testing.

Authentication in the backend is only enabled if `DEBUG=False` (the
default, can be set in .env file) is set.

The .env file in the frontend is not needed anymore since it only
contained the API KEY.

resolves #164 

## Changes

I removed the `api-key` command from the CLI.

In the backend I changed the settings in settings.py

In the frontend I included the cookies in the requests to the backend by
adding `credentials: "include",` to the fetch requests.

In backend and frontend I had to update the tests.

I updated the documentation of the
[CLI](https://github.com/brendel-group/mission_db/blob/feature/rest-api-session-auth/docs/cli/README.md),
the [Rest
API](https://github.com/brendel-group/mission_db/blob/feature/rest-api-session-auth/docs/restapi/README.md)
and the
[frontend](https://github.com/brendel-group/mission_db/blob/feature/rest-api-session-auth/docs/frontend/README.md)

## How to test

Login at the frontend

You can see the cookies and if they are sent in the developer tools in
the web storage tab under cookies and in the network tab when you click
on a request to the backend.

You can test that the requests fail when you manually delete the
`sessionid` cookie.

When you manually delete the `sessionid` cookie the session in the
backend is still stored for 2 weeks. You can delete all sessions in the
backend with SQL:
```sql
DELETE FROM django_session;
```

You can manually test the API Endpoints by trying to get missions with
[http://localhost:8000/restapi/missions/](http://localhost:8000/restapi/missions/)
and you should get a HTTP 403 response. If you then login using the
endpoint at
[http://localhost:8000/restapi/auth/login](http://localhost:8000/restapi/auth/login)
and try to get the missions again it should work.

Remember to log out after this with
[http://localhost:8000/restapi/auth/logout](http://localhost:8000/restapi/auth/logout).