CLI use django storage (#183)
The CLI now uses the default storage configured in settings.py for
filesystem operations.\
This means adding single missions with addfolder requires only the
relative path inside the folder backend/media.\
I renamed the `syncfolder` command to `sync` and removed the `--path`
argument, because it now automatically searches the backend/media folder
or when using AWS S3 buckets it will search there.

I also changed some arguments of some commands to be mutually exclusive.
For example the `tag remove` command had the optional arguments `--id`
and `--name`. They are now in a mutually exclusive group, which means
that only one of them is required but both are not allowed. This is
better because it is handled by `argparse` and not our own code.

I also changed the arguments when only the id is required so that it's
not required to use `--id` and you can just give the id. For example the
`mission remove` command previous:
```python
mission remove --id 1
```
vs now:
```python
mission remove 1
```

I updated the CLI documentation accordingly.

I fixed the tests that were failing previously and added a test for
adding details with the `addfolder` command.

resolves #180 

## How to test:
Move the example mission folders into the backend/media folder (or a
symbolic link) so that the folder structure is like this:
```
backend/media/
├── 2024.10.14_picking_apples
│   ├── test
│   │   └── bag_1728916594.8984177
│   └── train
│       ├── bag_1728916695.477267
│       └── bag_1728916777.3753355
└── 2024.10.15_picking_even_more_apples
    ├── test
    │   └── bag_1728916683.6689544
    └── train
        ├── bag_1728916621.5093482
        └── bag_1728916632.3249676
```
and use the `addfolder` or the now called `sync` command to add the
Missions and Files