DB Schema Topics (#169)
In this PR I added the required tables to support topics to the database
schema and also updated the models definition in models.py to directly
add these tables.

I added a allowlist for the topic names as a separate table and when a
topic is added it has to have a name in that table. Also the primary key
is the name so a new topic can be added without a lookup in the
allowed_topic_names table and django/postgres will check if it exists
like this:
```python
Topic.objects.create(name_id="topic/name", message_count=123, frequency=10.2, file=file, type='topic/type')
```
It will raise a `ForeignKeyViolation` if the topic type is not in the
Allowed_topic_types table.

To limit possible values it would also be possible to set `choices` for
the field, but that is supposed to be a static setting, while the
current approach with another table is dynamic.

It is also possible to prepopulate the `allowed_topic_names` table with
default values. Let me know if you think we should do this.

I also changed the `file_path` field of the `File` table to a
`FileField` and renamed it to file. Like this it now uses the django
storage api which not only supports local files but also AWS S3 buckets
or other storages via the package `django-storages` and I added the
`video` field to the `File` table.\
With django storages the files now have to be in a folder set by
`MEDIA_ROOT` (when using local filesystem) which is currently the new
folder `backend/media` and a file can be added using the relative path
to it in the `backend/media` folder. For example with
```python
File.objects.create(file="example/file.mcap", robot="Spot", duration=123, size=123)
```
(file is at `backend/media/example/file.mcap`)\
To not break the frontend which expects a `file_path` field I modified
the FileSerializer to contain a `file_path` field and not a `file`
field.