Implemented Rest API for topics (#203)
Hi, this resolves PR #166.

It adds a way to fetch Topics by a file path, which is unique to the
system. Topics are only allowed to have specific names, for this
purpose, GET, POST and DELETE requests for managing the allowed topic
names are added as well.

All new additions are fully tested and documented.

---------

Co-authored-by: Pascal Kuppler <75665115+pascal260303@users.noreply.github.com>