Converting MCAP files to mp4 (#154)
Hey there,

after our last meeting, @DerSimi told me, his biggest Christmas wish
would be a Python script, which converts `MCAP` files to their
respective `mp4` file. With this PR, I'm happily granting this wish 🎄

## What was changed
I created
[mcap_to_mp4.py](https://github.com/brendel-group/mission_db/blob/94bcafbe9f4a3837955a0c00176f5f8dd320beea/backend/mcap_to_mp4.py),
which is a Python script, that can extract the video data from a `MCAP`
file and converts said data to a `mp4` file. The connection with the CLI
will come in a separate PR

## How to test
The `mcap_file` variable defines the path to the `MCAP` file to be
converted. To test this function, you must replace the empty string with
a valid path. After running the script, a file called `output.mp4` will
be created, which contains the video file of the used `MCAP` file