Added basic pagination (#221)
Hi,
this PR resolves issue #219 by adding a pagination system to the mission
table. Currently, this is only frontend related, meaning no reduced
fetching.

![image](https://github.com/user-attachments/assets/2b5c0f34-4152-4279-ab95-786ca637b1d7)
On the bottom left, there's the option to specify the amount of entries
per page.


![image](https://github.com/user-attachments/assets/4e1f1cf0-3ea9-4e9f-991e-ae9587aa17bb)
The default value here is 10 entries per page. If desired, all elements
can be displayed at once. The setting is stored in a cookie:

![image](https://github.com/user-attachments/assets/803a863d-f7d5-448f-964e-acffc02ac079)

For this purpose, I added the` js-cookie` package, so don't forget to
run `npm install` before testing it.