Change the kind of relation of Mission and File (#218)
The relation between Missions and Files in the database was actually the
wrong kind of relation.
It was a many-to-many relation, but in the Filesystem because of the
Folder Structure it is a many-to-one relation.

A many-to-many relation would mean that a mission can have multiple
files (what we want) but a file can also belong to multiple missions
(not possible in the folder structure). With a many-to-one relation only
the first part is possible which fits our needs better.

I changed it so that it's a many-to-one relation.

The Mission_files table is deleted and the File table has the new fields
`mission_id` and `type`.

I changed everything from the CLI to the frontend, so that this change
doesn't break anything.

A File can not exist anymore without a Mission which it belongs to.

Everything should work like before.\
Only the REST API endpoint to get files by mission doesn't return a list
of jsons with `type` and `file` anymore with `file` being another json.
Now it just returns a list of jsons that contain the fields of the file
table (defined by `FileSerializer`).\

resolves #216 

## How to test
run
```bash
python manage.py makemigrations
python manage.py migrate
```
It may be required to provide a one-off default for the `mission_id` and
`type` fields.\
For the `mission_id` field pick a valid id of an existing mission and
for the type field you can use an empty string.\
Then delete all missions with the CLI and re-add them with the `sync`
command.

After that everything should work like before.