Sync files (#224)
resolves #108 

# Changes
- The sync_files function adds Files and their details to the database
if they are not in the database yet and it removes Files and their
details from the database if they are not in the filesystem anymore
- I added a custom logging handler that tracks if any log message was
emitted during the sync_folder command. If no log was emitted, that
means that nothing was changed/synced. Therefore the sync command now
has a message that tells the user that nothing was synced.

# Note
During refactoring I noticed that we use the yaml file to extract data.
In the future we want all extraction to happen from the mcap file. But
that is something for another Issue