Added type tags and filter option to the detail view. (#246)
Hi, this PR solves issue #245 by adding a new column to visualize the
dataset type (train or test). The type is identified by the prefix
train/ and test/ from the file path. If the prefix is unknown, the type
is replaced by ?.

The prefix train/ and test/ is now omitted in the file name, to prevent
visual redundancy.


![image](https://github.com/user-attachments/assets/f6d0ed71-a03c-42e5-8875-b12fa58e3dbe)

To filter for a type, click on a type tag, as previously implemented in
the Mission table.