Set same session age in backend as in frontend (#257)
The `MAX_SESSION_AGES` setting in the frontend and `SESSION_COOKIE_AGE`
in the backend are doing the same thing, but `SESSION_COOKIE_AGE` still
used the default value which is 2 weeks.\
This changes it to 1 hour like in the frontend.

Let me know if you think this is good.