Added dynamic type handling in the detail view (#258)
Hi, this PR resolves issue #251 by dynamically recognizing the parent
folder of a file.

For example:
successful_picks/bag_1737641283.2575762/bag_1737641283.2575762_0.mcap

successful_picks would be the type. If no type is present, the fallback
type is `?`.
The type is always recognized by checking if the mcap file does not
start with the same string as the top level folder.

Currently, the database does not provide colors for the type, so i
developed a deterministic method to always choose the same colors, it
works like this:
Given is a color map, a color index and a dictionary mapping type name
to color:
The first encountered type becomes the first color in the list, the next
the second, and so on. In case of an overflow, it begins in the front.


![image](https://github.com/user-attachments/assets/e3828624-be72-42bb-88b2-fedd283c2127)
In the image you see the current color list, that train appears two
times with the same color as expected and after a few types, red has to
appear again.


![image](https://github.com/user-attachments/assets/e1b6b871-82b2-4ce0-9686-fc4a17e67930)