Make it possible to change allowed topic names (#256)
I changed the table `allowed_topic_names` so that it can be used as a
blacklist of topic names and not a whitelist.\
Topics are checked by a validator if their name is in the now called
`Denied_topics` table.\
It's important to run `topic.full_clean()` when adding a new topic for
the validators to be executed.\
It's still possible to add topic names to the `Denied_topics` table with
the REST API like before with the `allowed_topic_names` table, but now
it's also possible to add, remove and list all denied topic names with
the CLI.

For this I added the new commands `topic allow <name>`, `topic deny
<name>` and `topic list-denied` \
I also added documentation and tests for them.

Denying a topic name means adding it the `Denied_topics` table and
removing existing topics with that name.\
Allowing a topic name means removing it from the `Denied_topics` table.\
Allowing a topic name that is not currently denied is not necessary and
will result in an error message.

resolves #250 

### How to test
You first need to run
```bash
python manage.py makemigrations
python manage.py migrate
```

Use the CLI to allow or deny any topic names.\
Currently you can't add topics via the sync command or other commands of
the CLI, but you could try manually adding a topic with a name in the
`Denied_topics` table. For this you can use `python manage.py shell` to
open a python shell where you don't have to do the django setup first.\
I also added a test case that tries to add a topic with a denied name,
so this is not really necessary.