Added Dataset View design (#261)
Hi this PR introduces a new dataset design and resolves issue #240. 

Old design:

![image](https://github.com/user-attachments/assets/0d2e6ac4-a7c8-43fe-84d9-a8a73b5e3a90)
New design:

![image](https://github.com/user-attachments/assets/336550a2-a79c-42de-948d-bf6effe8a016)

"OUTDATED" is type of the mcap file. 

There is also already space for the topics table @KageKrito will add.
Also added a new way to display the video absence:

![image](https://github.com/user-attachments/assets/2c7cac3a-6063-4f6c-a573-4c8d05eb1838)

This PR includes further design changes, for example, in the dataset
view the buttons where changed to a light design:

![image](https://github.com/user-attachments/assets/6d949648-ba08-40b4-8576-451721fb5e0c)

The same for the Tag Picker:

![image](https://github.com/user-attachments/assets/c82b3a02-a5fd-4848-8cf9-e8cad5e6025d)
Just for consistency.