Extended mp4 conversion to all topics with image message type (#262)
Hey there 👋,

this PR introduces the feature to generate a `mp4` file out of all
topics with the `sensor_msgs/msg/Image` message type (which is the only
one carrying data to generate images/videos) and would resolve the
second task of #207.

## How to test
In order to test this script manually, you have to modify the `bagpath`
variable in line 8 to the directory in which the `.mcap` file is
located. e.g `home\bag_1234\bag_1234_0.mcap` would turn into
`home\bag_1234`.
Simply running the script would then create the mp4 files and save them
in the directory previously defined in the `bagpath` variable. The
filename is orientated on the topic used to create the file, however
every `/` was replaced by a `-` due to the filenaming restrictions of
Windows

![image](https://github.com/user-attachments/assets/62ecd4d9-8729-4cfc-9b36-2f2eb86af65d)