Move video to topic table (#270)
The video field is now part of the topic table.\
I also updated the serializers, and views to match the changes.\
I didn't need to update any tests, because this was not explicitly
tested and also didn't add any tests for that.\
I also made sure the file_url is never `null`

## How to test
Manually add a topic with all fields and use any mp4 file as the video.
Use the restapi to get the topics of a file and test the access to the
video.