CLI command to generate videos (#274)
I integrated the way to extract/generate the videos from the mcap file
added by #262 into the CLI.\
It now has the command `generate-videos`.\
Because it was not possible to generate the videos directly in the
storage when something like S3 is used it now copies the mcap files to a
temporary folder (`backend/tmp/`) generates the videos there and moves
the videos to S3.\
When the mcap files are stored locally it works on that folder directly
without moving the files back and forth.

I also made it possible that the videos are stored locally and in a
different folder even if the mcap files are stored in S3.\
This can be controlled by the env var `STORE_VIDEO_LOCALLY`.

It also skips videos if it finds a file with the name for that video.

I also wrote documentation for the new CLI command also explaining this.

I did not write any tests, because it would require mocking many things
and then there is not much that is really tested.

resolves #265 

## How to test
Run the CLI command `generate-videos` on any mcap file that is already
in the database, it should create mp4 files for the videos in the same
folder as the mcap file.\
I also added autocompletion for the path (also works when stored in S3).