Datasetview video enhancement (#273)
Hi,
this PR implements a custom video view, with the ability to switch
between different topic videos of a single file.

What was changed:

![image](https://github.com/user-attachments/assets/50bd9639-baac-4750-98a1-34f7b07126a9)
Unnecessary control inputs such as volume, play back speed and video
download were removed and a new design was implemented. The arrows to
the left and right allow to switch the video by clicking them, at the
bottom left is the topic name related to the video displayed.


![image](https://github.com/user-attachments/assets/14d1faa9-74e4-4f0d-97af-8e52ee6ca0e2)

For time reasons, the design in the Fullscreen view is not overwritten. 

Currently, video fetching is not implemented using range requests, but
considering the minimal video length and resolution, this is not
considered a problem.