Determining message count and frequency of mcap topics (#264)
Hello again 👋,

this PR resolves issue #263 and implements functionality to calculate
the message count and the frequency of every topic.

## How to test
The best way to test the functionality is by running
`parse_rosbag_info(bagpath)` in the script and comparing the calculated
values with the ones Foxglove is calculating. Opening the `mcap` files
in Foxglove has become much easier since #260.