Fix video codec (#282)
The videos where generated with the `mp4v` codec which can't be played
in some browsers (for example firefox).\
I changed it to `h264/avc1` codec which is supported.

To use this it is necessary to install the python package
`opencv-python` not from the binary but to compile it. I adjusted the
requirements.txt so that `pip` does that but it may be necessary to
install some libraries in you OS. I used this
[guide](https://rockyshikoku.medium.com/use-h264-codec-with-cv2-videowriter-e00145ded181)
and the `Required build dependencies` section in the [official
documentation](https://docs.opencv.org/4.x/d2/de6/tutorial_py_setup_in_ubuntu.html)
for installation in Ubuntu.

The compilation by pip takes quite some time (github took ~27min)

I suspect in Ubuntu it is only necessary to install the required
dependencies in the official documentation and the packages `x264
libx264-dev`.\
It would be great if someone could test this before installing all.

I also changed the fps of the video to the frequency of that topic. With
this all videos of one mcap have the same length.

## How to test
Install libraries in your OS.
run
```python
pip install -r requirements.txt
```
generate some videos with the `generate-videos` command of the CLI and
open them in the browser. \
You can see the codec information in VLC under Tools > Codec Information